requirements.txt
naganami-mqtt==0.0.7
pyserial
